+++ RELEASENOTES.md
@@ -3,6 +3,7 @@ These are the release notes for the [mediawiki-api](README.md).
 ## Version 0.5 (development)
 
 * Introduces RevisionUndoer service
+* Introduces UserCreator service
 
 
 ## Version 0.4 (13 January 2015)